requirements.txt
pandas==2.2.2
numpy==1.26.4
MetaTrader5==5.0.45
pytz==2024.1
python-dateutil==2.9.0.post0
joblib==1.4.2
scikit-learn==1.5.2
xgboost==2.1.1
ta==0.11.0 # pure-Python technical indicators (no TA-Lib build hassles)
scipy==1.13.1
rich==13.7.1 # pretty console output
pyyaml==6.0.2